// lib/api/generated/model/sendMessageDtoType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API Documentation
 * The API description
 * OpenAPI spec version: 1.0
 */

export type SendMessageDtoType = typeof SendMessageDtoType[keyof typeof SendMessageDtoType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SendMessageDtoType = {
  TEXT: 'TEXT',
  SYSTEM: 'SYSTEM',
  STICKER: 'STICKER',
} as const;