// lib/api/generated/model/chatMessageDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API Documentation
 * The API description
 * OpenAPI spec version: 1.0
 */
import type { MessagePayloadDto } from './messagePayloadDto';

export interface ChatMessageDto {
  id: number;
  type: string;
  content: string;
  payload?: MessagePayloadDto;
  isFromUser: boolean;
  createdAt: string;
}