// lib/api/generated/model/sendMessageDtoOptionsItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API Documentation
 * The API description
 * OpenAPI spec version: 1.0
 */

export type SendMessageDtoOptionsItem = typeof SendMessageDtoOptionsItem[keyof typeof SendMessageDtoOptionsItem];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SendMessageDtoOptionsItem = {
  NEED_TRANSLATION: 'NEED_TRANSLATION',
  NEED_GRAMMAR_CORRECTION: 'NEED_GRAMMAR_CORRECTION',
} as const;